refactor(tips): dedupe getInitials/formatSharePct into tipDistribution util
TipTopEarners (new) and TipDistribution had byte-identical private
getInitials/formatSharePct helpers. Extract both into
src/utils/tipDistribution.ts, the shared aggregation module both
components already import from, and have each component import them
instead of maintaining its own copy.
