+++ src/components/tips/TipTopEarners.tsx
@@ -4,7 +4,12 @@ import { useMemo } from 'react';
 import { Button } from '@/components/ui/button';
 
 // Utils
-import { aggregateTipDistribution, type EmployeeDistribution } from '@/utils/tipDistribution';
+import {
+  aggregateTipDistribution,
+  formatSharePct,
+  getInitials,
+  type EmployeeDistribution,
+} from '@/utils/tipDistribution';
 import { formatCurrencyFromCents } from '@/utils/tipPooling';
 
 // Types
@@ -115,15 +120,4 @@ function TopEarnerRow({ employee }: { employee: EmployeeDistribution }) {
       </div>
     </li>
   );
-}
-
-function formatSharePct(pct: number): string {
-  return `${pct.toFixed(1)}%`;
-}
-
-function getInitials(name: string): string {
-  const parts = name.trim().split(/\s+/).filter(Boolean);
-  if (parts.length === 0) return '?';
-  if (parts.length === 1) return parts[0].slice(0, 2).toUpperCase();
-  return (parts[0][0] + parts[parts.length - 1][0]).toUpperCase();
 }
