+++ src/utils/tipDistribution.ts
@@ -159,3 +159,16 @@ export function paymentStatus(distribution: EmployeeDistribution): PaymentStatus
   if (unpaidCents === 0) return 'paid';
   return paidCents > 0 ? 'partial' : 'unpaid';
 }
+
+/** Formats an `EmployeeDistribution.sharePct` for display, e.g. `80.0%`. */
+export function formatSharePct(pct: number): string {
+  return `${pct.toFixed(1)}%`;
+}
+
+/** Derives up-to-two-letter avatar initials from a display name. */
+export function getInitials(name: string): string {
+  const parts = name.trim().split(/\s+/).filter(Boolean);
+  if (parts.length === 0) return '?';
+  if (parts.length === 1) return parts[0].slice(0, 2).toUpperCase();
+  return (parts[0][0] + parts[parts.length - 1][0]).toUpperCase();
+}
